Singularity: Sequentialize all table events
This avoids subtle reentrancy issues, especially with calculated
columns.

If an earlier RowAdded event handler changes a column value in the new
row (this happens for calculated columns), later handler groups would
end up seeing a ValueChanged event before the row's RowAdded event.
This broke the grid, which would add the row in response to the
ValueChanged for the new index, then add the row again in response to
the actual RowAdded event.